separacion de componentes
