+++ src/state/state.js
@@ -3,4 +3,8 @@ import { create } from 'zustand'
 export const storeBulkBuddies = create((set) => ({
   mobileMenuOpen: false,
   setMobileMenuOpen: (value) => set(() => ({ mobileMenuOpen: value })),
+  userState: {},
+  setUserState: (value) => set(() => ({ userState: value })),
+  isAuth: false,
+  setIsAuth: (value) => set(() => ({ isAuth: value })),
 }))
