+++ vite.config.js
@@ -1,8 +1,15 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react-swc'
+import path from "path";
 
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
+  resolve: {
+    alias: {
+      // eslint-disable-next-line no-undef
+      "@components": path.resolve(__dirname, "/src/components/*"),
+    },
+  },
 
 })
